Move ResourceHelper to class file
